Extract FunctionRegistry interface
Rename FunctionRegistry to GlobalFunctionRegistry
