+++ presto-main/src/test/java/com/facebook/presto/operator/aggregation/AbstractTestAggregationFunction.java
@@ -13,7 +13,7 @@
  */
 package com.facebook.presto.operator.aggregation;
 
-import com.facebook.presto.metadata.FunctionRegistry;
+import com.facebook.presto.metadata.GlobalFunctionRegistry;
 import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.block.BlockBuilder;
 import com.facebook.presto.spi.block.BlockBuilderStatus;
@@ -32,7 +32,7 @@ import static com.facebook.presto.type.UnknownType.UNKNOWN;
 
 public abstract class AbstractTestAggregationFunction
 {
-    protected final FunctionRegistry functionRegistry = new FunctionRegistry(new TypeRegistry(), true);
+    protected final GlobalFunctionRegistry functionRegistry = new GlobalFunctionRegistry(new TypeRegistry(), true);
     public abstract Block getSequenceBlock(int start, int length);
 
     protected final InternalAggregationFunction getFunction()
